Support authentication using the GitHub Personal Access Token (PAT).
GitHub limits unauthenticated requests to 60 per hour for each IP address.
This restriction can be avoided by authenticating with a PAT.

For security reasons, the PAT is not transmitted during proxy requests
and will not be stored long-term in settings.txt.

Additionally, FastGit is unavailable due to its service shutdown.
Therefore, it has been removed from the default proxy list.